Implemented vehicle damage reduction. Updated fuel handler and inv system
