+++ MSF/MSF_Functions.hpp
@@ -101,12 +101,14 @@ class CfgFunctions
 				recompile = 1;
 			};
 			class InfiniteFuelHandler {};
+			class SetVehicleDamageReduction {};
 		};
 		class Helpers
 		{
 			file = "msf\fnc\helpers";
 			class GetVehiclesByProperty {};
 			class GetPlayersByProperty {};
+			class GetVehiclesByPropertyValue {};
 		};
 		class Validators
 		{
