+++ MSF/cfg/MSF_Config.hpp
@@ -4,7 +4,7 @@ class MSFConfig
 	{
 		items[] = {"ACE_DefusalKit", "ACE_CableTie", "ToolKit", "ACE_UAVBattery", "ACE_wirecutter", "ACE_IR_Strobe_Item", "ACE_Yardage450"};
 		natoStdMags[] = {"100Rnd_65x39_caseless_black_mag", "100Rnd_65x39_caseless_khaki_mag", "100Rnd_65x39_caseless_mag", "100Rnd_65x39_caseless_black_mag_tracer", "ACE_100Rnd_65x39_caseless_mag_Tracer_Dim", "100Rnd_65x39_caseless_khaki_mag_tracer", "100Rnd_65x39_caseless_mag_Tracer", "200Rnd_65x39_cased_Box", "ACE_200Rnd_65x39_cased_Box_Tracer_Dim", "200Rnd_65x39_cased_Box_Tracer", "200Rnd_65x39_cased_Box_Red", "20Rnd_650x39_Cased_Mag_F", "30Rnd_65x39_caseless_black_mag", "30Rnd_65x39_caseless_green", "30Rnd_65x39_caseless_khaki_mag", "30Rnd_65x39_caseless_msbs_mag", "30Rnd_65x39_caseless_msbs_mag_Tracer", "30Rnd_65x39_caseless_mag", "30Rnd_65x39_caseless_green_mag_Tracer", "30Rnd_65x39_caseless_black_mag_Tracer", "ACE_30Rnd_65x39_caseless_green_mag_Tracer_Dim", "ACE_30Rnd_65x39_caseless_mag_Tracer_Dim", "30Rnd_65x39_caseless_khaki_mag_Tracer", "30Rnd_65x39_caseless_mag_Tracer", "ACE_20Rnd_65_Creedmor_mag", "ACE_30Rnd_65_Creedmor_black_mag", "ACE_30Rnd_65_Creedmor_khaki_mag", "ACE_30Rnd_65_Creedmor_msbs_mag", "ACE_30Rnd_65_Creedmor_mag", "200Rnd_65x39_cased_Box_Tracer_Red", "7Rnd_408_Mag", "ACE_7Rnd_408_305gr_Mag", "130Rnd_338_Mag"};
-		launcherRounds[] = {"MRAWS_HE_F", "MRAWS_HEAT55_F", "MRAWS_HEAT_F", "Titan_AA", "Titan_AT"};
+		launcherRounds[] = {"MRAWS_HEAT_F", "Titan_AA", "Titan_AT"};
 		grenades[] = {"HandGrenade", "SmokeShellGreen", "1Rnd_Smoke_Grenade_shell", "1Rnd_SmokeGreen_Grenade_shell", "1Rnd_SmokeBlue_Grenade_shell", "1Rnd_SmokeOrange_Grenade_shell", "1Rnd_SmokePurple_Grenade_shell", "1Rnd_SmokeYellow_Grenade_shell", "UGL_FlareGreen_F", "UGL_FlareCIR_F", "UGL_FlareRed_F", "UGL_FlareWhite_F", "ACE_M84", "ACE_M14", "1Rnd_HE_Grenade_shell"};		
 		aceMedical[] = {"ACE_bloodIV_500", "ACE_bloodIV", "ACE_quikclot", "ACE_packingBandage", "ACE_elasticBandage", "ACE_fieldDressing", "ACE_adenosine", "ACE_bodyBag", "ACE_bodyBag_blue", "ACE_bodyBag_white", "ACE_EarPlugs", "ACE_epinephrine", "ACE_morphine", "ACE_plasmaIV", "ACE_plasmaIV_500", "ACE_plasmaIV_250", "ACE_salineIV", "ACE_salineIV_250", "ACE_salineIV_500", "ACE_splint", "ACE_tourniquet"};
 		aceFood[] = {"ACE_Can_RedGull", "ACE_Can_Spirit", "ACE_Canteen", "ACE_Canteen_Empty", "ACE_Canteen_Half", "ACE_Humanitarian_Ration", "ACE_MRE_BeefStew", "ACE_MRE_ChickenTikkaMasala", "ACE_MRE_ChickenHerbDumplings", "ACE_MRE_CreamChickenSoup", "ACE_MRE_CreamTomatoSoup", "ACE_MRE_LambCurry", "ACE_MRE_MeatballsPasta", "ACE_MRE_SteakVegetables", "ACE_Sunflower_Seeds", "ACE_WaterBottle", "ACE_WaterBottle_Empty", "ACE_WaterBottle_Half", "ACE_Can_Franta"};
