+++ MSF/loadouts/Loadout_NATO.hpp
@@ -673,11 +673,10 @@ class TFY_ATRifleman
         "rhs_mag_30Rnd_556x45_M855A1_PMAG_Tracer_Red",
         "rhs_mag_30Rnd_556x45_M855A1_PMAG_Tracer_Red",
         "rhs_mag_30Rnd_556x45_M855A1_PMAG_Tracer_Red",
-        "MRAWS_HEAT55_F",
-        "MRAWS_HEAT55_F",
-        "MRAWS_HEAT55_F",
-        "MRAWS_HEAT55_F",
-        "MRAWS_HEAT55_F",
+        "MRAWS_HEAT_F",
+        "MRAWS_HEAT_F",
+        "MRAWS_HEAT_F",
+        "MRAWS_HEAT_F",
         "SmokeShellGreen",
         "SmokeShellGreen",
         "SmokeShellGreen",
