+++ MSF/loadouts/Loadouts_NATO_night.hpp
@@ -761,11 +761,10 @@ class TFY_ATRifleman_N
         "rhs_mag_30Rnd_556x45_M855A1_PMAG_Tracer_Red",
         "rhs_mag_30Rnd_556x45_M855A1_PMAG_Tracer_Red",
         "rhs_mag_30Rnd_556x45_M855A1_PMAG_Tracer_Red",
-        "MRAWS_HEAT55_F",
-        "MRAWS_HEAT55_F",
-        "MRAWS_HEAT55_F",
-        "MRAWS_HEAT55_F",
-        "MRAWS_HEAT55_F",
+        "MRAWS_HEAT_F",
+        "MRAWS_HEAT_F",
+        "MRAWS_HEAT_F",
+        "MRAWS_HEAT_F",        
         "SmokeShellGreen",
         "SmokeShellGreen",
         "SmokeShellGreen",
